Fixed Sum array for all types

+++ src/c/CudaMex/MexSumArray.cpp
@@ -17,35 +17,35 @@ void MexSumArray::execute( int nlhs, mxArray* plhs[], int nrhs, const mxArray* p
 		unsigned char* imageIn;
 		setupImagePointers(prhs[0],&imageIn,&imageDims);
 
-		sm = cSumArray(imageIn,imageDims.product(),device);
+		sm = (double)cSumArray(imageIn,imageDims.product(),device);
 	}
 	else if (mxIsUint16(prhs[0]))
 	{
 		unsigned short* imageIn;
 		setupImagePointers(prhs[0],&imageIn,&imageDims);
 
-		sm = cSumArray(imageIn,imageDims.product(),device);
+		sm = (double)cSumArray(imageIn,imageDims.product(),device);
 	}
 	else if (mxIsInt16(prhs[0]))
 	{
 		short* imageIn;
 		setupImagePointers(prhs[0],&imageIn,&imageDims);
 
-		sm = cSumArray(imageIn,imageDims.product(),device);
+		sm = (double)cSumArray(imageIn,imageDims.product(),device);
 	}
 	else if (mxIsUint32(prhs[0]))
 	{
 		unsigned int* imageIn;
 		setupImagePointers(prhs[0],&imageIn,&imageDims);
 
-		sm = cSumArray(imageIn,imageDims.product(),device);
+		sm = (double)cSumArray(imageIn,imageDims.product(),device);
 	}
 	else if (mxIsInt32(prhs[0]))
 	{
 		int* imageIn;
 		setupImagePointers(prhs[0],&imageIn,&imageDims);
 
-		sm = cSumArray(imageIn,imageDims.product(),device);
+		sm = (double)cSumArray(imageIn,imageDims.product(),device);
 	}
 	else if (mxIsSingle(prhs[0]))
 	{
